Create `ipynb` directory
It seems that the report generator script wants to write the notebooks into
the ipynb directory:

https://github.com/oscovida/oscovida.github.io/runs/4377991160?check_suite_focus=true#step:7:254

This is probably not required any more, as the ipynb directory is not used by
binder anylonger. But for now it breaks the daily build. So we recreate the
directory here, hoping that this will make the update of data pass again.